Add files via upload
Still using a First Person Controller.
I tried to hard code everything into the script so that it wouldn't break when I ported it over, and there would be less files. I was wondering if Juanita you could look at it and see and help me implement the AR Functionality on top of this? 
Two things! 1) The game is located inside the scene in the folder: Game 1 - Bone Locating Names Assets.
And 2) to make this look the way I was testing it, I used the resolution of 720x1280 to the right of Display, the third tab down, I set it to that, so it would look more like the phones these days. I also included a screen shot of what it is looking like.

Really excited to see how of if I can implement the AR from here!