Add files via upload
Adding return Bones functionality script to the project. 
You can add it to the bones and also can uncomment all the lines in return bones and switch warp move to "SlowlyMoveTowards" if you want to make it like woosh to the spot.
Ye! But ya like I did WarpBones first and I just thought to keep it for debugging purposes.