Made the questions match their answers
Made the questions match their answers, also, I added an animation that plays when the user clicks the correct and answer and next question.
From here, I will need someone to help me improve and modify the code so that a score is stored, using this code. 
And also make it so that the questions are randomized, similar to the way we did it earlier for the last game!